refactor: update contract and plan toolkits to use MAPBaseToolKit and improve tool validation methods
